+++ robotframework-ls/docs/changelog.md
@@ -1,4 +1,4 @@
-Next
+New in 1.3.5 (2022-10-19)
 -----------------------------
 
 ### Bugfixes
@@ -6,7 +6,9 @@ Next
 - If a user is paused in a log message break he should still be able to evaluate keywords based on the previous stack context.
 - Right click on folder and selecting `Robot Framework -> Run/Debug Tests/Tasks Suite` properly runs tests again. [#776](https://github.com/robocorp/robotframework-lsp/issues/776)
 - In the `interactive console`, if a multi-line statement is entered without a `***` section prefix, properly indent considering new lines.
+- Colors in the `interactive console` could be wrong on some cases.
 - Restore behavior that showed errors for keyword callers in peek error window (configurable through `robot.run.peekError.showErrorsInCallers`).
+- Fixed internal issue in debugger where a `path` variable could be accessed when undefined.
 
 
 New in 1.3.0 (2022-10-06)